[cli] temp: rename package - may revert
