+++ packages/backstopjs-config/README.md
@@ -1,11 +1,11 @@
 # Backstop JS Config
 
-This package is 
+This package provides configuration for visual regression testing with Backstop JS. 
 
 ## Usage
 
 ```
-const visualRegressionTesting = require('visual-regression-testing');
-
-// TODO: DEMONSTRATE API
+npm install @penskemediacorp/backstopjs-config --save-dev
 ```
+
+Once the package is installed, in larva.config.js, add the following object.
