+++ packages/backstopjs-config/lib/utils.js
